Task 3 finished. Task 2 still has to be checked
I have corrected Task 3 and now the field coordinates for each foot of each player are displayed + message of Static or Moving.
I believe that the threshold needs to be tuned a bit (choose a better value, but it's up for debate).

Task 2 it's still to be checked because the field coordinates don't seem ok. 

To do:
 - Check possible errors in the homography matrix or the field coordinates computation.
 - Check the code and the results to decide whether the results are ok, or not.

Tasks 4 and 5 to be started.

Thank you guys!